Request Equatable conformance to main module
