+++ JNetworkingKit/Source/Models/Request/Request.swift
@@ -3,7 +3,7 @@ import Foundation
 public typealias RequestHeaders = [String: String]
 public typealias RequestParameters = [String: String]
 
-public struct Request {
+public struct Request: Equatable {
     public enum Method: String {
         case get
         case post
@@ -52,5 +52,38 @@ extension Request: CustomStringConvertible {
                 "\n\troute = \(routeString)" +
                 "\n\tmethod = \(method)" +
                 "\n\theaders = \(headers)\n\tparameters = \(parameters)\n\tdata = \(dataString)>"
+    }
+}
+
+extension Request {
+    public static func == (lhs: Request, rhs: Request) -> Bool {
+        lhs.environment.url == rhs.environment.url
+            && lhs.route == rhs.route
+            && lhs.method == rhs.method
+            && lhs.headers == rhs.headers
+            && lhs.parameters == rhs.parameters
+            && compareDatas(lhsData: lhs.data, rhsData: rhs.data)
+    }
+
+    private static func compareDatas(lhsData: Data?, rhsData: Data?) -> Bool {
+        switch (lhsData, rhsData) {
+        case (.none, .none): return true
+        case (.none, .some): return false
+        case (.some, .none): return false
+        case (.some(let lhsData), .some(let rhsData)):
+            let lhsDictionary = NSDictionary(dictionary: parse(data: lhsData))
+            let rhsDictionary = parse(data: rhsData)
+
+            return lhsDictionary.isEqual(to: rhsDictionary)
+        }
+    }
+
+    private static func parse(data: Data) -> [String: Any] {
+        guard let parsed = try? JSONSerialization.jsonObject(with: data),
+              let dictionary = parsed as? [String: Any] else {
+            return [:]
+        }
+
+        return dictionary
     }
 }
